Fix file name.
